// src/model/user.ts
/**
 * Created by mehr30na on 7/12/17.
 */
import {UserType} from "./UserType";

export class User {

  id: number;
  firstName: string;
  lastName: string;
  username: string;
  password: string;
  userType: UserType;
  phoneNumber: string;
  email: string;
  token: string;
  image: string;
  day: number;
  month: number;
  year: number;

}